pyproject.toml
[tool.poetry]
name = "meshlium-sensor-filter"
version = "0.1.0"
description = ""
authors = ["Piotr Hajder <phajder@agh.edu.pl>"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector-python = "^8.0.26"
matplotlib = "^3.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
